fix: Rename carpets and route
