+++ src/components/ChannelHeader/styles.css
@@ -17,7 +17,7 @@ html {
   margin-right: 24px;
   flex-grow: 0;
 }
-.logo div {
+.logo img {
   background-color: #606060;
   width: 100%;
   height: 100%;
