+++ src/components/ChannelHeader/styles.scss
@@ -20,7 +20,7 @@ html{
     margin-right: 24px;
     flex-grow: 0;
 
-    div{
+    img{
         background-color: #606060;
         width: 100%;
         height: 100%;
