+++ src/components/PageManager/index.jsx
@@ -1,12 +1,14 @@
 import './styles.scss'
-import Channel_Header from '../Channel-Header'
-import TabCarousel from '../Tab-Carousel'
+import Channel_Header from '../ChannelHeader'
+import TabCarousel from '../TabCarousel'
 import VideoIntroduccion from '../VideoIntroduccion'
-import ListVideo from '../List-Video'
+import ListVideo from '../ListVideo'
+import Header from '../Header'
 
 function PageManager(){
     return(
         <div className="manager">
+                <Header/>
                 <Channel_Header/>
                 <TabCarousel/>
                 <VideoIntroduccion/>
