+++ src/components/TabCarousel/styles.scss
