+++ src/components/VideoItem/styles.scss
@@ -26,6 +26,8 @@
 .video-thumbnail{
     height: 118px;
     width: 210px;
-    background-color: #606060;
+    background-image: url("https://i.ytimg.com/vi/mZ22llUcnNE/hq720.jpg?sqp=-oaymwEcCNAFEJQDSFXyq4qpAw4IARUAAIhCGAFwAcABBg==&rs=AOn4CLCthRDBF0Md_xh1VRFiged5LaPqKQ");
+    background-repeat: repeat;
+    background-size: cover;
 }
